Edit Maven groupID/artifactID and change overall packaging
