+++ src/main/java/org/commonwl/view/CwlViewerApplication.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.commonwl.viewer;
+package org.commonwl.view;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
