+++ src/main/java/org/commonwl/view/researchobject/ROBundle.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.commonwl.viewer.researchobject;
+package org.commonwl.view.researchobject;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
@@ -26,8 +26,8 @@ import org.apache.taverna.robundle.Bundles;
 import org.apache.taverna.robundle.manifest.Agent;
 import org.apache.taverna.robundle.manifest.Manifest;
 import org.apache.taverna.robundle.manifest.PathMetadata;
-import org.commonwl.viewer.github.GitHubService;
-import org.commonwl.viewer.github.GithubDetails;
+import org.commonwl.view.github.GitHubService;
+import org.commonwl.view.github.GithubDetails;
 import org.eclipse.egit.github.core.CommitUser;
 import org.eclipse.egit.github.core.RepositoryCommit;
 import org.eclipse.egit.github.core.RepositoryContents;
