+++ src/main/java/org/commonwl/view/workflow/WorkflowController.java
@@ -17,11 +17,11 @@
  * under the License.
  */
 
-package org.commonwl.viewer.workflow;
+package org.commonwl.view.workflow;
 
 import org.apache.commons.lang.StringUtils;
-import org.commonwl.viewer.github.GithubDetails;
-import org.commonwl.viewer.graphviz.GraphVizService;
+import org.commonwl.view.github.GithubDetails;
+import org.commonwl.view.graphviz.GraphVizService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
